bug fix which tests get ran on each minter
