+++ test/minter-suite-minters/DA/MinterDALin/MinterDALinV2.test.ts
@@ -20,6 +20,7 @@ import {
 import { ONE_MINUTE, ONE_HOUR, ONE_DAY } from "../../../util/constants";
 import { MinterDALin_Common } from "./MinterDALin.common";
 import { MinterDAV1V2_Common } from "../MinterDAV1V2.common";
+import { MinterDAV2_Common } from "../MinterDAV2.common";
 
 /**
  * These tests intended to ensure this Filtered Minter integrates properly with
@@ -108,6 +109,10 @@ describe("MinterDALinV2_V3Core", async function () {
 
   describe("common DA V1V2 tests", async function () {
     MinterDAV1V2_Common();
+  });
+
+  describe("common DA V2 tests", async function () {
+    MinterDAV2_Common();
   });
 
   describe("setProjectMaxInvocations", async function () {
