+++ test/minter-suite-minters/MinterHolder/MinterHolder.common.ts
@@ -964,36 +964,6 @@ export const MinterHolder_Common = async () => {
     });
   });
 
-  describe("additional payee payments", async function () {
-    it("handles additional payee payments", async function () {
-      const valuesToUpdateTo = [
-        this.projectOne,
-        this.accounts.artist.address,
-        this.accounts.additional.address,
-        50,
-        this.accounts.additional2.address,
-        51,
-      ];
-      await this.genArt721Core
-        .connect(this.accounts.artist)
-        .proposeArtistPaymentAddressesAndSplits(...valuesToUpdateTo);
-      await this.genArt721Core
-        .connect(this.accounts.deployer)
-        .adminAcceptArtistAddressesAndSplits(...valuesToUpdateTo);
-
-      await this.minter
-        .connect(this.accounts.artist)
-        ["purchase(uint256,address,uint256)"](
-          this.projectZero,
-          this.genArt721Core.address,
-          this.projectZeroTokenZero.toNumber(),
-          {
-            value: this.pricePerTokenInWei,
-          }
-        );
-    });
-  });
-
   describe("setProjectMaxInvocations", async function () {
     it("handles getting tokenInfo invocation info with V1 core", async function () {
       await this.minter
