+++ test/minter-suite-minters/SetPrice/MinterSetPrice/MinterSetPriceV2.test.ts
@@ -20,6 +20,7 @@ import {
 
 import { MinterSetPrice_ETH_Common } from "./MinterSetPrice.common";
 import { MinterSetPriceV1V2_Common } from "../MinterSetPriceV1V2.common";
+import { MinterSetPriceV2_Common } from "../MinterSetPriceV2.common";
 
 /**
  * These tests intended to ensure this Filtered Minter integrates properly with
@@ -141,6 +142,10 @@ describe("MinterSetPriceV2_V3Core", async function () {
 
   describe("common MinterSetPrice V1V2 tests", async function () {
     MinterSetPriceV1V2_Common();
+  });
+
+  describe("common MinterSetPrice V2 tests", async function () {
+    MinterSetPriceV2_Common();
   });
 
   describe("setProjectMaxInvocations", async function () {
