+++ test/minter-suite-minters/SetPrice/MinterSetPriceV1V2.common.ts
@@ -42,39 +42,6 @@ export const MinterSetPriceV1V2_Common = async () => {
     });
   });
 
-  describe("purchase_H4M", async function () {
-    it("allows `purchase_H4M` by default", async function () {
-      await this.minter
-        .connect(this.accounts.user)
-        .purchase_H4M(this.projectZero, {
-          value: this.pricePerTokenInWei,
-        });
-    });
-  });
-
-  describe("additional payee payments", async function () {
-    it("handles additional payee payments", async function () {
-      const valuesToUpdateTo = [
-        this.projectZero,
-        this.accounts.artist2.address,
-        this.accounts.additional.address,
-        50,
-        this.accounts.additional2.address,
-        51,
-      ];
-      await this.genArt721Core
-        .connect(this.accounts.artist)
-        .proposeArtistPaymentAddressesAndSplits(...valuesToUpdateTo);
-      await this.genArt721Core
-        .connect(this.accounts.deployer)
-        .adminAcceptArtistAddressesAndSplits(...valuesToUpdateTo);
-
-      await this.minter.connect(this.accounts.user).purchase(this.projectZero, {
-        value: this.pricePerTokenInWei,
-      });
-    });
-  });
-
   describe("reentrancy attack", async function () {
     it("does not allow reentrant purchaseTo", async function () {
       // attacker deploys reentrancy contract
